Complete refactor to Redux
