+++ package.json
@@ -4,7 +4,6 @@
   "private": true,
   "homepage": "http://local.sivanandabahamas.org/price_calculator",
   "dependencies": {
-    "lodash": "^4.17.4",
     "material-ui": "^0.17.1",
     "moment": "^2.19.1",
     "prop-types": "^15.6.0",
